Inicializaciones de los vértices agregadas
En espera de las descripciones de Raúl para actualizar

+++ Cujae/src/logica/Facultad.java
@@ -9,7 +9,7 @@ public class Facultad extends LugarDeInteres{
 	
 	
 	
-	public Facultad(String id, float x, float y, String foto, String fotoReverso,String nombre, String descripcion, String anotaciones,String fotoPortada, String decano, String vicedecano) {
+	public Facultad(String id, double x, double y, String foto, String fotoReverso,String nombre, String descripcion, String anotaciones,String fotoPortada, String decano, String vicedecano) {
 		super(id, x, y, foto, fotoReverso, nombre, descripcion, anotaciones, fotoPortada);
 		
 		this.decano = decano;
